code and use cases for override fw_rules
